// src/components/CollaborativeTable/types.ts
export type UserPresence = {
    user_id: string;
    user_name: string;
    user_avatar?: string;
    online_at: string;
    last_activity_at: string;
};

export type CursorPosition = {
    user_id: string;
    row_id: string;
    column_id: string;
    x: number;
    y: number;
    timestamp: string;
};